Use type `panel` for toolbar button in Pale Moon
This makes the button look consistent in Pale Moon on Windows, see #257. You can use the hidden setting `toolbarButtonForceTypeMenu` to revert to the previous design.

This also sets the button's `flip` attribute to` both` to improve the automatic positioning of the panel when the browser is not in full screen mode.